backend/player_photo_generator/source/requirements.txt
functions-framework==3.*

firebase-admin

google-cloud-aiplatform
pillow

opencv-python
numpy
tensorflow
scipy
